Added cross platform support
