+++ example/onfido_web_sdk_example.dart
@@ -1,6 +1,2 @@
-import 'package:onfido_web_sdk/onfido_web_sdk.dart';
 
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}
+void main() {}
