+++ test/onfido_web_sdk_test.dart
@@ -1,16 +1,3 @@
-import 'package:onfido_web_sdk/onfido_web_sdk.dart';
-import 'package:test/test.dart';
 
 void main() {
-  group('A group of tests', () {
-    final awesome = Awesome();
-
-    setUp(() {
-      // Additional setup goes here.
-    });
-
-    test('First Test', () {
-      expect(awesome.isAwesome, isTrue);
-    });
-  });
 }
